Add before and after specs
