+++ cypress/integration/plugins/observability-dashboards/8_after.spec.js
@@ -1,0 +1,14 @@
+/*
+ * Copyright OpenSearch Contributors
+ * SPDX-License-Identifier: Apache-2.0
+ */
+
+/// <reference types="cypress" />
+
+describe('After', () => {
+  before(() => {
+    cy.deleteAllIndices();
+  });
+
+  it('clean up complete', () => {});
+});
